pkg/fms: Move system config to pkg/fms/system
