+++ pkg/fms/system/system_setup.go
@@ -1,4 +1,4 @@
-package fms
+package system
 
 import (
 	"bytes"
