+++ pkg/fms/system/tpl/coresvc.sh.tpl
