+++ pkg/fms/system/tpl/grafana_dash_gizmo.json.tpl
