+++ pkg/fms/system/tpl/grafana_dashboards.yaml.tpl
