+++ pkg/fms/system/tpl/icewm_startup.tpl
