+++ pkg/fms/system/tpl/icewm_theme.tpl
