+++ pkg/fms/system/tpl/prometheus.yml.tpl
