+++ pkg/fms/system/tpl/xorg.conf.tpl
